requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.5.3
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==46.0.3
Django==4.2.27
django-appconf==1.0.6
django-celery-beat==2.8.1
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-timezone-field==7.2.1
idna==3.11
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary @ file:///C:/Users/User/Desktop/avito_bidder_project/psycopg2_binary-2.9.9-cp38-cp38-win_amd64.whl#sha256=a6cdcc3ede532f4a4b96000b6362099591ab4a3e913d70bcbac2b56c872446f7
pycparser==2.23
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.1.1
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.5
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.14
